requirements.txt
beautifulsoup4==4.12.2
gTTS==2.3.2
nltk==3.8.1
numpy==1.23.5
openai==0.27.7
python-dotenv==1.0.0
Requests==2.31.0
textdescriptives==2.6.1
